Moved CompressedPakResource and its deps to new cmpd.py file
